requirements.txt
streamlit
pandas
requests
beautifulsoup4
xlsxwriter
python-dotenv
openpyxl